+++ compiler-tests/src/test/data/diagnostic/dependencygraph/AsContributionCannotWorkOnContributedExtensions.kt
@@ -22,9 +22,8 @@ interface FooProvider {
   val foo: Foo
 }
 
-fun box(): String {
+fun example() {
   val appGraph = createGraph<AppGraph>()
   val childGraph = appGraph.asContribution<ChildGraph.Factory>().create()
   val foo = <!AS_CONTRIBUTION_ERROR!>childGraph<!>.asContribution<FooProvider>().foo
-  return "OK"
 }